% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opart_gaussian.R
\name{opart_gaussian}
\alias{opart_gaussian}
\title{compute the optimal changepoint model for a vector of real-valued data
and a non-negative real-valued penalty,
given the square loss (to minimize) / gaussian likelihood (to maximize)}
\usage{
opart_gaussian(n_data, data_ptr, penalty)
}
\arguments{
\item{n_data}{A number indicating the number of data points}

\item{data_ptr}{A list of numbers for which the changepoint model is to be computed}

\item{penalty}{A non-negative real number indicating penalty}
}
\value{
A pointer to the optimal cost values and a pointer to the optimal segment ends
}
\description{
compute the optimal changepoint model for a vector of real-valued data
and a non-negative real-valued penalty,
given the square loss (to minimize) / gaussian likelihood (to maximize)
}
\examples{
opart_gaussian(3, c(1,2,3), 1)
opart_gaussian(4, c(1,2,3,4), 1)
}
